1394/05/04

+++ app/src/main/java/com/ariana/shahre_ma/WebServiceGet/HTTPGetAreaJosn.java
@@ -76,27 +76,26 @@ public class HTTPGetAreaJosn extends AsyncTask<String,Void,Integer>
      */
         @Override
         protected void onPostExecute(Integer result) {
-          //  onPostExecute(result);
-            if(len_Area>0)
-            {
-                try
-                {
-                    DataBaseSqlite db=new DataBaseSqlite(context);
-                    db.delete_Area();
-                    for (int i = 0; i < len_Area; i++) {
-                        db.Add_area(Id_area[i],Name_Area[i],CityId_area[i]);
-                    }
 
-                    pd.dismiss();
-                } catch (Exception e)
-                {
-                    pd.dismiss();
-                }
-            }
-            else
+          //  onPostExecute(result);
+            if(result==1)
             {
-                pd.dismiss();
-                //pToast.makeText(getApplicationContext(),"همگام سازی انجام نشد دوباره امتحان کنید",Toast.LENGTH_LONG).show();
+                if (len_Area > 0) {
+                    try {
+                        DataBaseSqlite db = new DataBaseSqlite(context);
+                        db.delete_Area();
+                        for (int i = 0; i < len_Area; i++) {
+                            db.Add_area(Id_area[i], Name_Area[i], CityId_area[i]);
+                        }
+
+                        pd.dismiss();
+                    } catch (Exception e) {
+                        pd.dismiss();
+                    }
+                } else {
+                    pd.dismiss();
+                    //pToast.makeText(getApplicationContext(),"همگام سازی انجام نشد دوباره امتحان کنید",Toast.LENGTH_LONG).show();
+                }
             }
         }
